wxH/vipMangementSystem/requirements.txt
altgraph==0.15
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.4
baidu-aip==2.1.0.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
DBUtils==1.2
dis3==0.1.2
Django==1.11.13
django-cors-headers==2.3.0
enum34==1.1.6
Flask==1.0.2
future==0.16.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.1
macholib==1.9
MarkupSafe==1.0
#matplotlib==2.1.1
#mysql-connector-python==1.0.12
#MySQL-python==1.2.3
numpy==1.13.3
olefile==0.44
opencv-python==3.4.0.12
parsel==1.4.0
pefile==2017.11.5
Pillow==4.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyInstaller==3.3.1
PyJWT==1.6.4
pymongo==3.6.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
#pypiwin32==219
python-dateutil==2.6.1
pytz==2017.3
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.5.0
